25.01.2018

+++ C_Sharp_WareHouse_Forms/CustomerDataGridView.cs
@@ -43,7 +43,6 @@ namespace C_Sharp_WareHouse_Forms
 
         private void CustomerDataGridView_Load(object sender, EventArgs e)
         {
-            //dataGridView1.Rows.Clear();
             DataTable table  = ConvertToDataTable(Containers.CustomerList);
             dataGridView1.DataSource = table;
         }
@@ -75,6 +74,12 @@ namespace C_Sharp_WareHouse_Forms
                 txtBoxEmailUpd.Text = dataGridView1.SelectedRows[0].Cells[3].Value.ToString();
                 maskdTxtBxPhoneUpd.Text = dataGridView1.SelectedRows[0].Cells[4].Value.ToString();
                 richTxtBxAddrUpd.Text = dataGridView1.SelectedRows[0].Cells[5].Value.ToString();
+                btnAdd.Enabled = false;
+                btnDelete.Enabled = false;
+                txtboxNameSearch.Enabled = false;
+                txtBxSurnameSearch.Enabled = false;
+                txtBxEmailSearch.Enabled = false;
+                msktxtBxSearch.Enabled = false;
                 dataGridView1.Enabled = false;
             }
             catch
@@ -86,13 +91,22 @@ namespace C_Sharp_WareHouse_Forms
 
         private void btnDelete_Click(object sender, EventArgs e)
         {
-            int id = Convert.ToInt32(dataGridView1.SelectedRows[0].Cells[0].Value);
-            if (MessageBox.Show("Are you sure to remove selected item?", "Delete", MessageBoxButtons.YesNo, MessageBoxIcon.Warning) == DialogResult.Yes)
+            try
             {
-                containers.DeleteCustomer(id);
-                MessageBox.Show("Removed", "Success", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                if (dataGridView1.SelectedRows is null || dataGridView1.SelectedRows[0].Cells[0].Value == null)
+                    throw new Exception("No any element to remove");
+                int id = Convert.ToInt32(dataGridView1.SelectedRows[0].Cells[0].Value);
+                if (MessageBox.Show("Are you sure to remove selected item?", "Delete", MessageBoxButtons.YesNo, MessageBoxIcon.Warning) == DialogResult.Yes)
+                {
+                    containers.DeleteCustomer(id);
+                    MessageBox.Show("Removed", "Success", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                }
+                return;
             }
-            return;
+            catch
+            {
+                MessageBox.Show("No any Customer Selected", "ERROR", MessageBoxButtons.OK, MessageBoxIcon.Error);
+            }
         }
 
         private void CustomerDataGridView_FormClosed(object sender, FormClosedEventArgs e)
@@ -104,22 +118,12 @@ namespace C_Sharp_WareHouse_Forms
         {
             pnlUpdate.Visible = false;
             dataGridView1.Enabled = true;
-        }
-
-        private void textBox3_KeyUp(object sender, KeyEventArgs e)
-        {
-            //Controls.OfType<TextBox>().FirstOrDefault(x => x.Text != string.Empty);
-            //string keyword =
-            //var matchedItems = (!int.TryParse(textBoxSearch.Text, out int digit) ?
-            //     Containers.ProductList.FindAll(item => item.ArticleName.IndexOf
-            //        (textBoxSearch.Text, 0, StringComparison.OrdinalIgnoreCase) >= 0 || item.Description.IndexOf
-            //        (textBoxSearch.Text, 0, StringComparison.OrdinalIgnoreCase) >= 0).ToList() :
-            //        Containers.ProductList.FindAll(item => item.UniqueID == Convert.ToInt32(textBoxSearch.Text)).ToList());
-
-            //dataGridView1.Rows.Clear();
-            DataTable table = ConvertToDataTable(Containers.CustomerList);
-            dataGridView1.DataSource = table;
-
+            btnAdd.Enabled = true;
+            btnDelete.Enabled = true;
+            txtboxNameSearch.Enabled = true;
+            txtBxSurnameSearch.Enabled = true;
+            txtBxEmailSearch.Enabled = true;
+            msktxtBxSearch.Enabled = true;
         }
 
         public static DataTable ConvertToDataTable<T>(IList<T> data)
@@ -139,6 +143,88 @@ namespace C_Sharp_WareHouse_Forms
             return table;
         }
 
+        private void txtboxNameSearch_Leave(object sender, EventArgs e)
+        {
+            if (txtboxNameSearch.Text == "")
+            {
+                txtboxNameSearch.Text = "Search by Name";
+                txtboxNameSearch.ForeColor = Color.Silver;
+            }
+        }
 
+        private void txtBxSurnameSearch_Leave(object sender, EventArgs e)
+        {
+            if (txtBxSurnameSearch.Text == "")
+            {
+                txtBxSurnameSearch.ForeColor = Color.Silver;
+                txtBxSurnameSearch.Text = "Search by Surname";
+            }
+
+        }
+
+        private void txtBxEmailSearch_Leave(object sender, EventArgs e)
+        {
+            if (txtBxEmailSearch.Text == "")
+            {
+                txtBxEmailSearch.ForeColor = Color.Silver;
+                txtBxEmailSearch.Text = "Search by Email";
+            }
+        }
+
+        private void txtboxNameSearch_KeyUp(object sender, KeyEventArgs e)
+        {
+            string keyword = txtboxNameSearch.Text;
+            var matchedItems = (int.TryParse(keyword, out int digit) ?
+                 throw new Exception("Please type only letters"):
+                 Containers.CustomerList.FindAll(item => item.FirstName.IndexOf
+                 (keyword, 0,StringComparison.OrdinalIgnoreCase) == 0)).ToList();
+            DataTable table = ConvertToDataTable(matchedItems);
+            dataGridView1.DataSource = table;
+        }
+
+        private void txtBxSurnameSearch_KeyUp(object sender, KeyEventArgs e)
+        {
+            string keyword = txtBxSurnameSearch.Text;
+            var matchedItems = (int.TryParse(keyword, out int digit) ?
+                 throw new Exception("Please type only letters") :
+                 Containers.CustomerList.FindAll(item => item.LastName.IndexOf
+                 (keyword, 0, StringComparison.OrdinalIgnoreCase) == 0)).ToList();
+            DataTable table = ConvertToDataTable(matchedItems);
+            dataGridView1.DataSource = table;
+        }
+
+        private void txtBxEmailSearch_KeyUp(object sender, KeyEventArgs e)
+        {
+            string keyword = txtBxEmailSearch.Text;
+            var matchedItems = (int.TryParse(keyword, out int digit) ?
+                 throw new Exception("Please type only letters") :
+                 Containers.CustomerList.FindAll(item => item.Email.IndexOf
+                 (keyword, 0, StringComparison.OrdinalIgnoreCase) == 0)).ToList();
+            DataTable table = ConvertToDataTable(matchedItems);
+            dataGridView1.DataSource = table;
+        }
+
+        private void ClearText(object sender, EventArgs e)
+        {
+            var txtBxSearch = (sender as TextBox);
+            if (txtBxSearch.Text.Contains("Search by "))
+            {
+                txtBxSearch.Text = "";
+                txtBxSearch.ForeColor = Color.Black;
+            }
+        }
+
+        private void msktxtBxSearch_KeyUp(object sender, KeyEventArgs e)
+        {
+            var matchedItems = Containers.CustomerList.FindAll(item => item.Contact.IndexOf
+            (msktxtBxSearch.Text, 0, StringComparison.OrdinalIgnoreCase) == 0).ToList();
+            DataTable table = ConvertToDataTable(matchedItems);
+            dataGridView1.DataSource = table;
+        }
+
+        private void msktxtBxSearch_Enter(object sender, EventArgs e)
+        {
+            msktxtBxSearch.ForeColor = Color.Black;
+        }
     }
 }
